feat(ui): extract ProjectStats and swap header for a sparkline under flag
- Move ProjectStats into its own file and rename the Relay fragment to
  ProjectStats_project (queryName: ProjectStatsQuery), updating all
  spread sites in ProjectPage, SpansTable, and SpanAnnotationsEditor.
- Add ProjectSpanCountSparkline, a minimal stacked-bar chart backed by
  traceCountByStatusTimeSeries with axes/grid/legend/tooltip hidden.
- When the tracing_ux flag is enabled, the project header renders the
  sparkline in place of the horizontal stats strip; otherwise the
  existing stats row is unchanged.
